+++ app/layout.tsx
@@ -1,10 +1,10 @@
 import "@/styles/globals.css";
-import SessionProvider from "@/components/provider/SessionProvider";
+import SessionProvider from "@/components/Provider/SessionProvider";
 import { unstable_getServerSession } from "next-auth/next";
 import { authOptions } from "@/lib/auth";
 import { Noto_Sans_SC } from "@next/font/google";
-import { Providers } from "@/components/provider/ThemeProvider";
-import { StateProvider } from "@/components/provider/StateProvider";
+import { Providers } from "@/components/Provider/ThemeProvider";
+import { StateProvider } from "@/components/Provider/StateProvider";
 import { AnalyticsWrapper } from "@/components/analytics";
 
 const notoSansSC = Noto_Sans_SC({
