+++ components/Card/CharacterAvatarCard.tsx
