+++ components/Card/PersonAvatarCard.tsx
