+++ components/DropdownMenu/MoreDropdownMenu.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { MoreDropdownMenu as MoreDropdownMenuComponent } from "@/components/Panel/Subject/MoreDropdownMenu";
+import { MoreDropdownMenu as MoreDropdownMenuComponent } from "@/components/DropdownMenu/MoreDropdownMenu";
 import { STORYBOOK_SUBJECT_ID } from "@/lib/constant";
 import { userEvent, screen, within } from "@storybook/testing-library";
 import { expect } from "@storybook/jest";
