+++ components/Select/CollectionTypeSelect.stories.tsx
@@ -1,20 +1,20 @@
-import type { Meta, StoryObj } from "@storybook/react";
-import { RatingSelect as RatingSelectComponent } from "@/components/Panel/Subject/RatingSelect";
-import { STORYBOOK_SUBJECT_ID } from "@/lib/constant";
-import { screen, userEvent, waitFor, within } from "@storybook/testing-library";
 import { expect } from "@storybook/jest";
+import type { Meta, StoryObj } from "@storybook/react";
+import { CollectionTypeSelect as CollectionTypeSelectComponent } from "@/components/Select/CollectionTypeSelect";
+import { STORYBOOK_SUBJECT_ID } from "@/lib/constant";
+import { userEvent, within, screen, waitFor } from "@storybook/testing-library";
 import { reactQueryDevtoolsDecorator } from "@/lib/storybook";
 
-const meta: Meta<typeof RatingSelectComponent> = {
-  title: "Select/RatingSelect",
-  component: RatingSelectComponent,
+const meta: Meta<typeof CollectionTypeSelectComponent> = {
+  title: "Select/CollectionTypeSelect",
+  component: CollectionTypeSelectComponent,
   decorators: [reactQueryDevtoolsDecorator],
 };
 
 export default meta;
-type Story = StoryObj<typeof RatingSelectComponent>;
+type Story = StoryObj<typeof CollectionTypeSelectComponent>;
 
-export const RatingSelect: Story = {
+export const CollectionTypeSelect: Story = {
   args: {
     subject_id: STORYBOOK_SUBJECT_ID,
   },
@@ -28,7 +28,7 @@ export const RatingSelect: Story = {
         await userEvent.click(selectEl);
         const listbox = within(screen.getByRole("listbox"));
         const options = await listbox.findAllByRole("option");
-        await expect(options.length).toEqual(11);
+        await expect(options.length).toEqual(5);
         for (const option of options) {
           await expect(option).toHaveTextContent(/./);
         }
