+++ components/Skeleton/SubjectContentSkeleton.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { SubjectContentSkeleton as SubjectContentSkeletonComponent } from "@/components/Panel/Subject/SubjectContentSkeleton";
+import { SubjectContentSkeleton as SubjectContentSkeletonComponent } from "@/components/Skeleton/SubjectContentSkeleton";
 
 const meta: Meta<typeof SubjectContentSkeletonComponent> = {
   title: "SubjectContentSkeleton",
